requirements.txt
azure-cosmos>=4.3.0
python-dotenv>=0.21.0
tqdm>=4.65.0
colorama>=0.4.6
